Rely on the new Y2Network::NtpServer class
