+++ src/lib/y2caasp/clients/kubic_minion_role_dialog.rb
@@ -23,6 +23,7 @@ require "cwm/dialog"
 require "y2caasp/widgets/kubic_admin_node"
 require "y2caasp/widgets/ntp_server"
 require "y2caasp/dhcp_ntp_servers"
+require "y2network/ntp_server"
 
 module Y2Caasp
   # This library provides a simple dialog for setting
@@ -57,45 +58,6 @@ module Y2Caasp
             Y2Caasp::Widgets::NtpServer.new(ntp_servers)
           ))
       )
-    end
-
-  private
-
-    #
-    # Propose the NTP servers from the DHCP response, fallback to a random
-    # machine from the ntp.org pool if enabled in control.xml.
-    #
-    # @return [Array<String>] proposed NTP servers, empty if nothing suitable found
-    #
-    def ntp_servers
-      # TODO: use Yast::NtpClient.ntp_conf if configured
-      # to better handle going back
-      servers = dhcp_ntp_servers
-      servers = ntp_fallback if servers.empty?
-
-      servers
-    end
-
-    #
-    # The fallback servers for NTP configuration
-    #
-    # @return [Array<String>] the fallback servers, empty if disabled in control.xml
-    #
-    def ntp_fallback
-      # propose the fallback when enabled in control file
-      return [] unless Yast::ProductFeatures.GetBooleanFeature("globals", "default_ntp_setup")
-
-      # copied from timezone/dialogs.rb:
-      base_products = Yast::Product.FindBaseProducts
-      host = if base_products.any? { |p| p["name"] =~ /openSUSE/i }
-        "opensuse"
-      else
-        # TODO: use a SUSE server when available in the future
-        "novell"
-      end
-
-      # propose a random pool server in range 0..3
-      ["#{rand(4)}.#{host}.pool.ntp.org"]
     end
   end
 end
